:sparkles: feature: customizable colors (#1977)
* :sparkles: feature: customizable colors

* :sparkles: feature: customizable colors

* :sparkles: feature: customizable colors
