+++ middleware/logger/logger.go
@@ -61,19 +61,6 @@ const (
 	TagReset      = "reset"
 )
 
-// Color values
-const (
-	cBlack   = "\u001b[90m"
-	cRed     = "\u001b[91m"
-	cGreen   = "\u001b[92m"
-	cYellow  = "\u001b[93m"
-	cBlue    = "\u001b[94m"
-	cMagenta = "\u001b[95m"
-	cCyan    = "\u001b[96m"
-	cWhite   = "\u001b[97m"
-	cReset   = "\u001b[0m"
-)
-
 // New creates a new middleware handler
 func New(config ...Config) fiber.Handler {
 	// Set default config
@@ -133,6 +120,9 @@ func New(config ...Config) fiber.Handler {
 			return c.Next()
 		}
 
+		// Alias colors
+		colors := c.App().Config().ColorScheme
+
 		// Set error handler once
 		once.Do(func() {
 			// get longested possible path
@@ -179,16 +169,16 @@ func New(config ...Config) fiber.Handler {
 			// Format error if exist
 			formatErr := ""
 			if chainErr != nil {
-				formatErr = cRed + " | " + chainErr.Error() + cReset
+				formatErr = colors.Red + " | " + chainErr.Error() + colors.Reset
 			}
 
 			// Format log to buffer
 			_, _ = buf.WriteString(fmt.Sprintf("%s |%s %3d %s| %7v | %15s |%s %-7s %s| %-"+errPaddingStr+"s %s\n",
 				timestamp.Load().(string),
-				statusColor(c.Response().StatusCode()), c.Response().StatusCode(), cReset,
+				statusColor(c.Response().StatusCode(), colors), c.Response().StatusCode(), colors.Reset,
 				stop.Sub(start).Round(time.Millisecond),
 				c.IP(),
-				methodColor(c.Method()), c.Method(), cReset,
+				methodColor(c.Method(), colors), c.Method(), colors.Reset,
 				c.Path(),
 				formatErr,
 			))
@@ -240,7 +230,7 @@ func New(config ...Config) fiber.Handler {
 				return buf.WriteString(c.Route().Path)
 			case TagStatus:
 				if cfg.enableColors {
-					return buf.WriteString(fmt.Sprintf("%s %3d %s", statusColor(c.Response().StatusCode()), c.Response().StatusCode(), cReset))
+					return buf.WriteString(fmt.Sprintf("%s %3d %s", statusColor(c.Response().StatusCode(), colors), c.Response().StatusCode(), colors.Reset))
 				}
 				return appendInt(buf, c.Response().StatusCode())
 			case TagResBody:
@@ -255,27 +245,27 @@ func New(config ...Config) fiber.Handler {
 				return buf.WriteString(c.Request().URI().QueryArgs().String())
 			case TagMethod:
 				if cfg.enableColors {
-					return buf.WriteString(fmt.Sprintf("%s %-7s %s", methodColor(c.Method()), c.Method(), cReset))
+					return buf.WriteString(fmt.Sprintf("%s %-7s %s", methodColor(c.Method(), colors), c.Method(), colors.Reset))
 				}
 				return buf.WriteString(c.Method())
 			case TagBlack:
-				return buf.WriteString(cBlack)
+				return buf.WriteString(colors.Black)
 			case TagRed:
-				return buf.WriteString(cRed)
+				return buf.WriteString(colors.Red)
 			case TagGreen:
-				return buf.WriteString(cGreen)
+				return buf.WriteString(colors.Green)
 			case TagYellow:
-				return buf.WriteString(cYellow)
+				return buf.WriteString(colors.Yellow)
 			case TagBlue:
-				return buf.WriteString(cBlue)
+				return buf.WriteString(colors.Blue)
 			case TagMagenta:
-				return buf.WriteString(cMagenta)
+				return buf.WriteString(colors.Magenta)
 			case TagCyan:
-				return buf.WriteString(cCyan)
+				return buf.WriteString(colors.Cyan)
 			case TagWhite:
-				return buf.WriteString(cWhite)
+				return buf.WriteString(colors.White)
 			case TagReset:
-				return buf.WriteString(cReset)
+				return buf.WriteString(colors.Reset)
 			case TagError:
 				if chainErr != nil {
 					return buf.WriteString(chainErr.Error())
